+++ moli-layout/Cargo.toml
@@ -2,22 +2,19 @@
 name = "moli-layout"
 version = "0.1.0"
 edition = "2024"
-license = "(MIT OR Apache-2.0) AND BSD-3-Clause"
+license.workspace = true
 
 [dependencies]
 euclid = "=0.22.14"
 kurbo = "=0.13.1"
 moli-image = { path = "../moli-image" }
+moli-system-fonts = { path = "../moli-system-fonts" }
 parley = "=0.11.1"
-parking_lot = "0.12"
 stylo = "0.20"
 stylo_taffy = { git = "https://github.com/DioxusLabs/blitz", rev = "5081c65811a4396f5a99b2e0aca542a4a4a6606f", default-features = false, features = ["std", "block", "flexbox", "grid", "floats"] }
 taffy = { version = "=0.13.0", default-features = false, features = ["std", "block_layout", "flexbox", "grid", "float_layout", "content_size", "calc", "detailed_layout_info"] }
 thiserror = "2.0.17"
 wuff = "=0.2.5"
-
-[target.'cfg(any(target_os = "linux", target_os = "freebsd"))'.dependencies]
-yeslogic-fontconfig-sys = "=6.0.1"
 
 [lints]
 workspace = true
