+++ moli-system-fonts/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "moli-system-fonts"
+version = "0.1.0"
+edition = "2024"
+license = "(MIT OR Apache-2.0) AND BSD-3-Clause"
+
+[dependencies]
+fontique = "=0.11.1"
+
+[target.'cfg(any(target_os = "linux", target_os = "freebsd"))'.dependencies]
+parking_lot = "0.12"
+yeslogic-fontconfig-sys = "=6.0.1"
+
+[lints]
+workspace = true
